% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/chooseModule.R
\name{chooseModule}
\alias{chooseModule}
\title{Choosing dense modules}
\usage{
chooseModule(res.list, top = 0.01, plot = FALSE)
}
\arguments{
\item{res.list}{The results from function \code{\link{dms}}}

\item{top}{Either a percentage (<1) or an integer (>=1)}

\item{plot}{To plot the sub-network or not}
}
\value{
A list contains both the chosen modules and the sub-network constructed from
the chosen modules.
}
\description{
This function will choose the top modules specified by users for downstream
analyses. Sub-network can also be plotted if \code{plot=TRUE}.
}
\details{
The parameter \code{top} could be either a percentage (<1) or an integer
(>=1). For example, if \code{top} is 0.01, the top 1\% modules will be
chosen; if \code{top} is 50, the top 50 modules will be chosen.
}
\examples{
\dontrun{
res.list <- dms(network, geneweight, expr1, expr2)
}
}
\seealso{
\code{\link{dms}}
}

