% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generate_edge_weight.R
\name{generate_edge_weight}
\alias{generate_edge_weight}
\title{Compute edge weight}
\usage{
generate_edge_weight(expr1, expr2, network, geneweight)
}
\arguments{
\item{expr1}{A data frame containing gene expression data from case samples.
The first column is gene identifier (should be coordinate with the node
symbol used in PPI}

\item{expr2}{A data frame containing gene expression data from control
samples. The first column should be the same as expr1}

\item{network}{A data frame containing a symbolic edge list of the PPI
network}

\item{geneweight}{A data frame containing two columns: the first is unique
gene identifier (should be coordinate with the node symbol used in PPI);
the second is gene-based p-value derived from GWAS}
}
\value{
A co-expression matrix is returned
}
\description{
\code{\link{generate_edge_weight}} computes differential gene co-expression
(i.e., the change of gene co-expression between case and control samples) to
infer the edge weight of PPI network. \code{\link{generate_edge_weight}} is
generally called by \code{\link{generate_graph}}. Typically users do not need
to call it.
}
\examples{
\dontrun{
edgeweight <- generate_edge_weight(expr1, expr2, network, geneweight)
}
}

