% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generate_graph.R
\name{generate_graph}
\alias{generate_graph}
\title{Build background PPI Network}
\usage{
generate_graph(expr1, expr2 = NULL, network, geneweight)
}
\arguments{
\item{expr1}{A data frame containing gene expression data from case samples.
The first column is gene identifier (should be coordinate with the node
symbol used in PPI}

\item{expr2}{A data frame containing gene expression data from control
samples. The first column should be the same as expr1}

\item{network}{A data frame containing a symbolic edge list of the PPI
network}

\item{geneweight}{A data frame containing two columns: the first is unique
gene identifier (should be coordinate with the node symbol used in PPI);
the second is gene-based p-value derived from GWAS}
}
\description{
\code{\link{generate_graph}} constructs a node- and edge-weighted PPI network
for function \code{\link{dms}}. This function is generally called by function
\code{\link{dms}}. Typically users do not need to call it.
}
\examples{
\dontrun{
G<- generate_graph (expr1, expr2, network, geneweight)
}
}

